Twitter_FactomV2/Streamer/requirements.txt
tweepy==3.7.0
factom-api==0.1.1
ipython
dataset
textblob
python-identitykeys
psycopg2>=2.7,<3.0